scripts/requirements.txt
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
